@@ -33,14 +33,14 @@
   "dependencies": {
     "bcrypt-nodejs": "^0.0.3",
     "body-parser": "^1.18.2",
-    "client-sessions": "^0.7.0",
+    "client-sessions": "^0.8.0",
     "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.3",
     "dotenv": "^4.0.0",
     "ejs": "^2.5.7",
     "express": "^4.16.1",
     "express-ejs-layouts": "^2.3.1",
-    "express-validator": "^2.21.0",
+    "express-validator": "^4.2.1",
     "leaflet.markercluster": "^1.1.0",
     "lunr": "^2.1.3",
     "mapbox.js": "^3.1.1",
@@ -49,14 +49,15 @@
     "morgan": "^1.9.0",
     "multer": "^1.3.0",
     "nodemailer": "^4.1.3",
-    "passport": "^0.3.2",
+    "passport": "^0.4.0",
     "passport-local": "^1.0.0",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-motion": "^0.4.8",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-motion": "^0.5.2",
     "react-onclickout": "^2.0.7",
     "request": "^2.83.0",
-    "sharp": "^0.17.3",
+    "sharp": "^0.18.4",
     "underscore": "^1.8.3",
     "universal-cookie": "^2.1.0"
   },
@@ -69,14 +70,14 @@
     "babelify": "^7.3.0",
     "browserify": "^14.4.0",
     "cssnano": "^3.10.0",
-    "eslint": "^3.19.0",
-    "eslint-plugin-react": "^6.10.3",
-    "mocha": "^3.5.3",
+    "eslint": "^4.8.0",
+    "eslint-plugin-react": "^7.4.0",
+    "mocha": "^4.0.1",
     "node-sass": "^4.5.3",
     "nodemon": "^1.12.1",
     "postcss-cli": "^4.1.1",
-    "react-router": "^3.2.0",
-    "should": "^11.2.1",
+    "react-router": "^4.2.0",
+    "should": "^13.1.0",
     "uglify-js": "^3.1.3"
   }
 }
